Fix bug related to cmd.Wait()
Instead of blocking with cmd.Wait(), now the goroutine is blocking with
cmd.Process.Wait().

cmd.Wait() waits not only for the process to exit, but also for any I/O
in the process' stdout and stderr to be closed. Since cmd is using a
CmdLogger as its stdout and stderr, cmd.Wait() just hangs. This still
needs a little more studying, but for now the presented solution should
be enough.

Go repo issue related to this: https://github.com/golang/go/issues/18874